feat(AI): add separate @mieweb/ui/ozwell entry point
Move OzwellWidget, OzwellAnimatedButton, and useOzwell exports out of
the main AI barrel into a dedicated src/ozwell.ts entry point, matching
the existing ag-grid and datavis pattern.

- Create src/ozwell.ts with all Ozwell exports
- Add 'ozwell' entry to tsup.config.ts
- Add './ozwell' to package.json exports map (ESM + CJS + types)
- Remove Ozwell re-exports from src/components/AI/index.ts
- Add 'Optional Add-ons' section to README documenting all three
  separate entry points (AG Grid, DataVis, Ozwell)

Consumers now import from '@mieweb/ui/ozwell' instead of '@mieweb/ui',
keeping @ozwell/react and @rive-app/react-canvas out of the core bundle.
