+++ README.md
@@ -47,6 +47,24 @@ This library requires React 18+ and React DOM 18+:
 npm install react react-dom
 ```
 
+### Optional Add-ons
+
+Heavy or specialized dependencies are kept in separate entry points so they don't bloat the core bundle. Install the peer dependencies for the add-ons you need:
+
+| Entry point              | Install                                              | Import path              |
+| ------------------------ | ---------------------------------------------------- | ------------------------ |
+| **AG Grid**              | `npm install ag-grid-community ag-grid-react`        | `@mieweb/ui/ag-grid`    |
+| **DataVis**              | `npm install datavis-ace`                             | `@mieweb/ui/datavis`    |
+| **Ozwell AI Widget**     | `npm install @ozwell/react`                           | `@mieweb/ui/ozwell`     |
+| Ozwell + animated button | `npm install @ozwell/react @rive-app/react-canvas`   | `@mieweb/ui/ozwell`     |
+
+```tsx
+// Example: Ozwell widget
+import { OzwellWidget } from '@mieweb/ui/ozwell';
+
+<OzwellWidget apiKey="agnt_key-your-key" />
+```
+
 ## Quick Start
 
 ### Option 1: With Tailwind CSS (Recommended)
